requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.0.4
Django==5.2.5
django-allauth==65.11.1
eval_type_backport==0.2.2
Flask==2.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==4.8.3
invoke==2.2.0
itsdangerous==2.0.1
Jinja2==3.0.3
lxml==6.0.1
MarkupSafe==2.0.1
mistralai==1.9.9
protobuf==3.19.6
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
setuptools==67.6.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions>=4.5,<5  # updated to satisfy anyio >=4.5 requirement
urllib3==2.5.0
Werkzeug==2.0.3
wheel==0.40.0
whitenoise==6.9.0
zipp==3.6.0

# packages used in code but previously missing
openpyxl
python-pptx
reportlab
